requirements.txt
anticaptchaofficial==1.0.24
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.1
mysql-connector-python==8.0.20
pdfminer==20191125
protobuf==3.11.3
py==1.8.1
pycryptodome==3.9.7
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
